requirements.txt
google-cloud-pubsub==2.18.4
loguru==0.7.2
pydantic==2.4.2
python-dotenv==1.0.0
sentry-sdk==1.31.0